viewport: consistently use floats for viewports
OpenGL < 4.0 only supports integer based viewports and internally we
have a mixture of code using floats and integers for viewports. This
patch switches all viewports throughout clutter and cogl to be
represented using floats considering that in the future we may want to
take advantage of floating point viewports with modern hardware/drivers.

+++ clutter/clutter-actor.c
@@ -1968,7 +1968,7 @@ _clutter_actor_fully_transform_vertices (ClutterActor *self,
   ClutterActor *stage;
   CoglMatrix modelview;
   CoglMatrix projection;
-  int viewport[4];
+  float viewport[4];
 
   g_return_val_if_fail (CLUTTER_IS_ACTOR (self), FALSE);
 
@@ -11786,7 +11786,7 @@ clutter_actor_get_paint_box (ClutterActor    *self,
   const ClutterPaintVolume *pv;
   CoglMatrix modelview;
   CoglMatrix projection;
-  int viewport[4];
+  float viewport[4];
   ClutterPaintVolume projected_pv;
 
   g_return_val_if_fail (CLUTTER_IS_ACTOR (self), FALSE);
