+++ clutter/clutter-stage-private.h
@@ -41,15 +41,15 @@ void                _clutter_stage_get_projection_matrix (ClutterStage          
                                                           CoglMatrix            *projection);
 void                _clutter_stage_dirty_projection      (ClutterStage          *stage);
 void                _clutter_stage_set_viewport          (ClutterStage          *stage,
-                                                          int                    x,
-                                                          int                    y,
-                                                          int                    width,
-                                                          int                    height);
+                                                          float                  x,
+                                                          float                  y,
+                                                          float                  width,
+                                                          float                  height);
 void                _clutter_stage_get_viewport          (ClutterStage          *stage,
-                                                          int                   *x,
-                                                          int                   *y,
-                                                          int                   *width,
-                                                          int                   *height);
+                                                          float                 *x,
+                                                          float                 *y,
+                                                          float                 *width,
+                                                          float                 *height);
 void                _clutter_stage_dirty_viewport        (ClutterStage          *stage);
 void                _clutter_stage_maybe_setup_viewport  (ClutterStage          *stage);
 void                _clutter_stage_maybe_relayout        (ClutterActor          *stage);
